// src/API.ts
export type Todo = {
  __typename: "Todo",
  id: string,
  description: string,
  date?: string | null,
  priority?: string | null,
  isCompleted?: boolean | null,
  createdAt: string,
  updatedAt: string,
};